Made a house, made a tree, modified the game so we put things on the surface.

+++ 3dTesting/MainWindow.xaml.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Windows;
 using System.Windows.Controls;
@@ -102,8 +101,8 @@ namespace _3dTesting
             }
 
             GameHelpers.UpdateMapOverlay(mapOverlay, surfaceMapBitmap,
-                Convert.ToInt16(world.WorldInhabitants.FirstOrDefault(z => z.ObjectName == "Surface")?.ParentSurface?.GlobalMapPosition.x),
-                Convert.ToInt16(world.WorldInhabitants.FirstOrDefault(z => z.ObjectName == "Surface")?.ParentSurface?.GlobalMapPosition.z));
+                Convert.ToInt32(world.WorldInhabitants.FirstOrDefault(z => z.ObjectName == "Surface")?.ParentSurface?.GlobalMapPosition.x),
+                Convert.ToInt32(world.WorldInhabitants.FirstOrDefault(z => z.ObjectName == "Surface")?.ParentSurface?.GlobalMapPosition.z));
 
             Task.Run(() =>
             {
